Move admin-only endpoints into metabase/admin/settings/api
api-key, email, ldap, google, oauth, slack, and cloud-migration are only
consumed from the admin module, so they move next to their consumers and
inject into the shared Api instance, following the geojson precedent.
email's api object is renamed emailApi to stop shadowing settings.ts's
settingsApi.
