+++ frontend/src/metabase/admin/settings/api/api-key.ts
@@ -1,3 +1,11 @@
+
+import { Api } from "metabase/api";
+import {
+  idTag,
+  invalidateTags,
+  listTag,
+  provideApiKeyListTags,
+} from "metabase/api/tags";
 import type {
   ApiKey,
   ApiKeyId,
@@ -7,9 +15,6 @@ import type {
   UpdateApiKeyRequest,
   UpdateApiKeyResponse,
 } from "metabase-types/api/admin";
-
-import { Api } from "./api";
-import { idTag, invalidateTags, listTag, provideApiKeyListTags } from "./tags";
 
 export const apiKeyApi = Api.injectEndpoints({
   endpoints: (builder) => ({
