+++ frontend/src/metabase/admin/settings/api/google.ts
@@ -1,7 +1,7 @@
-import type { EnterpriseSettings } from "metabase-types/api";
 
-import { Api } from "./api";
-import { invalidateTags, tag } from "./tags";
+import { Api } from "metabase/api";
+import { invalidateTags, tag } from "metabase/api/tags";
+import type { EnterpriseSettings } from "metabase-types/api";
 
 type GoogleAuthSettings = Pick<
   EnterpriseSettings,
