+++ frontend/src/metabase/admin/settings/api/slack.ts
@@ -1,7 +1,7 @@
-import type { SlackAppInfo, SlackSettings } from "metabase-types/api";
 
-import { Api } from "./api";
-import { listTag } from "./tags";
+import { Api } from "metabase/api";
+import { listTag } from "metabase/api/tags";
+import type { SlackAppInfo, SlackSettings } from "metabase-types/api";
 
 export const slackApi = Api.injectEndpoints({
   endpoints: (builder) => ({
