+++ frontend/src/metabase/admin/settings/components/ApiKeys/CreateApiKeyModal.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from "react";
 import { t } from "ttag";
 
-import { useCreateApiKeyMutation } from "metabase/api";
+import { useCreateApiKeyMutation } from "metabase/admin/settings/api/api-key";
 import {
   Form,
   FormErrorMessage,
