+++ frontend/src/metabase/admin/settings/components/CloudPanel/MigrationInProgress.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable ttag/no-module-declaration -- see metabase#55045 */
 import { t } from "ttag";
 
-import { useCancelCloudMigrationMutation } from "metabase/api";
+import { useCancelCloudMigrationMutation } from "metabase/admin/settings/api/cloud-migration";
 import { ExternalLink } from "metabase/common/components/ExternalLink";
 import { useToggle } from "metabase/common/hooks/use-toggle";
 import type { Plan } from "metabase/common/utils/plan";
