+++ frontend/src/metabase/admin/settings/components/Email/SelfHostedSMTPConnectionForm.tsx
@@ -3,7 +3,7 @@ import { useCallback } from "react";
 import {
   useDeleteEmailSMTPSettingsMutation,
   useUpdateEmailSMTPSettingsMutation,
-} from "metabase/api/email";
+} from "metabase/admin/settings/api/email";
 
 import { BaseSMTPConnectionForm } from "./BaseSMTPConnectionForm";
 import { trackSMTPSetupSuccess } from "./analytics";
