+++ frontend/src/metabase/api/index.ts
@@ -3,14 +3,12 @@ export * from "./activity";
 export * from "./ai-entity-analysis";
 export * from "./analytics";
 export * from "./api";
-export * from "./api-key";
 export * from "./automagic-dashboards";
 export * from "./bookmark";
 export * from "./bug-report";
 export * from "./cache-config";
 export * from "./card";
 export * from "./channel";
-export * from "./cloud-migration";
 export * from "./collection";
 export * from "./comment";
 export * from "./dashboard";
@@ -23,10 +21,8 @@ export * from "./exploration";
 export * from "./field";
 export * from "./frontend-errors";
 export * from "./glossary";
-export * from "./google";
 export * from "./group-table-access-policy";
 export * from "./index-manager";
-export * from "./ldap";
 export * from "./llm";
 export * from "./login-history";
 export * from "./measure";
@@ -35,7 +31,6 @@ export * from "./metric";
 export * from "./model-index";
 export * from "./moderation";
 export * from "./notification";
-export * from "./oauth";
 export * from "./parameters";
 export * from "./permission";
 export * from "./persist";
@@ -49,7 +44,6 @@ export * from "./security-center";
 export * from "./segment";
 export * from "./session";
 export * from "./setup";
-export * from "./slack";
 export * from "./snippet";
 export * from "./subscription";
 export * from "./table";
